Split monster code
