+++ tower-defense/src/game_state/monster/mod.rs
@@ -1,9 +1,25 @@
+mod monster_hp_bar;
+mod monster_template;
+mod move_monsters;
+mod render;
 mod skill;
 
-use super::{user_status_effect::UserStatusEffectKind, *};
+use crate::{
+    MapCoordF32,
+    game_state::projectile::ProjectileTargetIndicator,
+    route::{MoveOnRoute, Route},
+};
+pub use monster_template::MonsterTemplate;
+pub use move_monsters::move_monsters;
 use namui::*;
-pub use skill::*;
-use std::sync::atomic::{AtomicUsize, Ordering};
+pub use skill::{
+    MonsterSkill, MonsterSkillTemplate, MonsterStatusEffect, MonsterStatusEffectKind,
+    PrebuiltSkill, activate_monster_skills, remove_monster_finished_status_effects,
+};
+use std::sync::{
+    Arc,
+    atomic::{AtomicUsize, Ordering},
+};
 
 const MONSTER_HP_BAR_HEIGHT: Px = px(4.);
 
@@ -77,184 +93,6 @@ impl Monster {
         self.id
     }
 }
-impl Component for &Monster {
-    fn render(self, ctx: &RenderCtx) {
-        // TODO: add monster image
-        let monster_wh = TILE_PX_SIZE * 0.6;
-        let path = Path::new().add_oval(Rect::from_xy_wh(monster_wh.to_xy() * -0.5, monster_wh));
-        let paint = Paint::new(Color::RED);
-        ctx.translate(TILE_PX_SIZE.to_xy() * 0.5)
-            .add(namui::path(path, paint));
-
-        let hp_bar_wh = Wh::new(monster_wh.width, MONSTER_HP_BAR_HEIGHT);
-        ctx.translate(Xy::new(
-            TILE_PX_SIZE.width * 0.5,
-            TILE_PX_SIZE.width * 0.5 + monster_wh.height * 0.6,
-        ))
-        .add(MonsterHpBar {
-            wh: hp_bar_wh,
-            progress: self.hp / self.max_hp,
-        });
-    }
-}
-
-struct MonsterHpBar {
-    wh: Wh<Px>,
-    progress: f32,
-}
-impl Component for MonsterHpBar {
-    fn render(self, ctx: &RenderCtx) {
-        let Self { wh, progress } = self;
-
-        let container_rect = Rect::from_xy_wh(wh.to_xy() * -0.5, wh);
-
-        ctx.add(rect(RectParam {
-            rect: Rect::from_xy_wh(container_rect.xy(), Wh::new(wh.width * progress, wh.height)),
-            style: RectStyle {
-                stroke: None,
-                fill: Some(RectFill { color: Color::RED }),
-                round: None,
-            },
-        }));
-
-        ctx.add(rect(RectParam {
-            rect: container_rect,
-            style: RectStyle {
-                stroke: Some(RectStroke {
-                    color: palette::OUTLINE,
-                    width: 1.px(),
-                    border_position: BorderPosition::Outside,
-                }),
-                fill: Some(RectFill {
-                    color: palette::SURFACE_CONTAINER,
-                }),
-                round: None,
-            },
-        }));
-    }
-}
-
-pub struct MonsterTemplate {
-    pub kind: MonsterKind,
-    pub max_hp: f32,
-    pub skills: Vec<MonsterSkillTemplate>,
-    pub velocity: Velocity,
-    pub damage: f32,
-    pub reward: usize,
-}
-impl MonsterTemplate {
-    fn velocity(mul: f32) -> Velocity {
-        Per::new(5.0 * mul, Duration::from_secs(1))
-    }
-    fn damage(mul: f32) -> f32 {
-        mul
-    }
-    fn reward(mul: usize) -> usize {
-        mul
-    }
-    pub fn new(kind: MonsterKind) -> Self {
-        let (max_hp, velocity, damage, reward, skills) = match kind {
-            MonsterKind::Mob01 => (10.0, 0.5, 1.0, 3, vec![]),
-            MonsterKind::Mob02 => (25.0, 0.5, 1.0, 3, vec![]),
-            MonsterKind::Mob03 => (90.0, 0.3, 1.0, 3, vec![]),
-            MonsterKind::Mob04 => (75.0, 1.1, 1.0, 3, vec![]),
-            MonsterKind::Mob05 => (250.0, 1.0, 1.0, 3, vec![PrebuiltSkill::Invincible01]),
-            MonsterKind::Mob06 => (850.0, 0.75, 1.0, 5, vec![]),
-            MonsterKind::Mob07 => (1750.0, 0.75, 1.0, 5, vec![]),
-            MonsterKind::Mob08 => (6000.0, 0.4, 1.0, 5, vec![]),
-            MonsterKind::Mob09 => (3500.0, 1.25, 1.0, 5, vec![PrebuiltSkill::Speedmul01]),
-            MonsterKind::Mob10 => (7500.0, 0.75, 1.0, 5, vec![PrebuiltSkill::Speedmul02]),
-            MonsterKind::Mob11 => (10000.0, 1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob12 => (15000.0, 1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob13 => (45000.0, 0.5, 1.0, 5, vec![]),
-            MonsterKind::Mob14 => (20000.0, 1.5, 1.0, 5, vec![PrebuiltSkill::ImmuneSlow01]),
-            MonsterKind::Mob15 => (45000.0, 1.0, 1.0, 5, vec![PrebuiltSkill::Invincible02]),
-            MonsterKind::Named01 => (100.0, 0.5, 3.0, 30, vec![PrebuiltSkill::ImmuneSlow02]),
-            MonsterKind::Named02 => (500.0, 0.75, 3.0, 30, vec![PrebuiltSkill::Invincible03]),
-            MonsterKind::Named03 => (1500.0, 1.0, 3.0, 30, vec![PrebuiltSkill::Speedmul03]),
-            MonsterKind::Named04 => (3500.0, 1.25, 3.0, 30, vec![PrebuiltSkill::ImmuneSlow03]),
-            MonsterKind::Named05 => (15000.0, 0.5, 5.0, 30, vec![PrebuiltSkill::Heal01]),
-            MonsterKind::Named06 => (30000.0, 0.75, 5.0, 30, vec![PrebuiltSkill::ImmuneSlow04]),
-            MonsterKind::Named07 => (45000.0, 1.0, 5.0, 30, vec![PrebuiltSkill::Heal02]),
-            MonsterKind::Named08 => (65000.0, 1.25, 5.0, 30, vec![PrebuiltSkill::Speedmul04]),
-            MonsterKind::Named09 => (200000.0, 0.5, 7.0, 30, vec![PrebuiltSkill::Invincible04]),
-            MonsterKind::Named10 => (250000.0, 0.75, 7.0, 30, vec![PrebuiltSkill::Heal03]),
-            MonsterKind::Named11 => (300000.0, 1.0, 7.0, 30, vec![PrebuiltSkill::Heal04]),
-            MonsterKind::Named12 => (300000.0, 1.25, 7.0, 30, vec![PrebuiltSkill::AreaSpeedmul01]),
-            MonsterKind::Named13 => (
-                650000.0,
-                0.5,
-                10.0,
-                30,
-                vec![PrebuiltSkill::AreaImmuneSlow01],
-            ),
-            MonsterKind::Named14 => (
-                550000.0,
-                0.75,
-                10.0,
-                30,
-                vec![PrebuiltSkill::AreaInvincible01],
-            ),
-            MonsterKind::Named15 => (
-                550000.0,
-                1.0,
-                10.0,
-                30,
-                vec![PrebuiltSkill::AreaInvincible02],
-            ),
-            MonsterKind::Named16 => (
-                750000.0,
-                1.25,
-                10.0,
-                30,
-                vec![PrebuiltSkill::AreaImmuneSlow02],
-            ),
-            MonsterKind::Boss01 => (3500.0, 1.0, 15.0, 50, vec![PrebuiltSkill::AreaInvincible03]),
-            MonsterKind::Boss02 => (10000.0, 1.0, 20.0, 75, vec![PrebuiltSkill::AreaHeal01]),
-            MonsterKind::Boss03 => (35000.0, 1.0, 20.0, 100, vec![PrebuiltSkill::AreaSpeedmul02]),
-            MonsterKind::Boss04 => (85000.0, 1.0, 25.0, 100, vec![PrebuiltSkill::AreaSpeedmul03]),
-            MonsterKind::Boss05 => (
-                200000.0,
-                1.0,
-                25.0,
-                125,
-                vec![PrebuiltSkill::AreaImmuneSlow03],
-            ),
-            MonsterKind::Boss06 => (300000.0, 1.0, 25.0, 125, vec![PrebuiltSkill::AreaHeal02]),
-            MonsterKind::Boss07 => (
-                375000.0,
-                1.0,
-                50.0,
-                125,
-                vec![PrebuiltSkill::AreaImmuneSlow04],
-            ),
-            MonsterKind::Boss08 => (500000.0, 1.0, 50.0, 125, vec![PrebuiltSkill::AreaHeal03]),
-            MonsterKind::Boss09 => (
-                700000.0,
-                1.0,
-                50.0,
-                125,
-                vec![PrebuiltSkill::AreaSpeedmul04],
-            ),
-            MonsterKind::Boss10 => (
-                850000.0,
-                1.0,
-                50.0,
-                125,
-                vec![PrebuiltSkill::AreaInvincible04],
-            ),
-            MonsterKind::Boss11 => (1125000.0, 1.0, 50.0, 125, vec![PrebuiltSkill::AreaHeal04]),
-        };
-        Self {
-            kind,
-            max_hp,
-            skills: skills.into_iter().map(|prebuilt| prebuilt.into()).collect(),
-            velocity: Self::velocity(velocity),
-            damage: Self::damage(damage),
-            reward: Self::reward(reward),
-        }
-    }
-}
 
 #[derive(Clone, Copy)]
 pub enum MonsterKind {
@@ -300,54 +138,4 @@ pub enum MonsterKind {
     Boss09,
     Boss10,
     Boss11,
-}
-
-pub fn move_monsters(game_state: &mut GameState, dt: Duration) {
-    for monster in &mut game_state.monsters {
-        let is_immune_to_slow = monster.status_effects.iter().any(|status_effect| {
-            matches!(status_effect.kind, MonsterStatusEffectKind::ImmuneToSlow)
-        });
-        let mut dt = dt;
-
-        for status_effect in &monster.status_effects {
-            match status_effect.kind {
-                MonsterStatusEffectKind::SpeedMul { mul } => {
-                    if is_immune_to_slow && mul < 1.0 {
-                        continue;
-                    }
-                    dt *= mul;
-                }
-                MonsterStatusEffectKind::Invincible | MonsterStatusEffectKind::ImmuneToSlow => {}
-            }
-        }
-
-        monster.move_on_route.move_by(dt);
-    }
-
-    let mut damage = 0.0;
-    game_state.monsters.retain(|monster| {
-        if monster.move_on_route.is_finished() {
-            damage += monster.get_damage_to_user();
-            return false;
-        }
-        true
-    });
-
-    for user_status_effect in &game_state.user_status_effects {
-        match user_status_effect.kind {
-            UserStatusEffectKind::DamageReduction { damage_multiply } => {
-                damage *= damage_multiply;
-            }
-        }
-    }
-
-    if game_state.shield > 0.0 {
-        let min = damage.min(game_state.shield);
-        damage -= min;
-        game_state.shield -= min;
-    }
-    game_state.hp -= damage;
-    if game_state.hp <= 0.0 {
-        game_state.goto_result();
-    }
 }
